More documentation cleanup/removal of Rects/Surfs
Removed rect attributes from most sprites - not useful when trying to blit stuff to specific positions. Rather, it feels more understandable to simply calculate positions within the surface I am bliting to, instead of the position being inherent to a sprite.
Added portal image to portal. Changed some instances of 'image' with 'image_name' to be more explicit, when image_name is used to just refer to the file_name of the image.