Reorganisation of Sidebar, AttackButton class.
Created AttackButton class instead of Sidebar handling the attack buttons.
Changed update() methods. Now Sidebar update() handles the entire sidebar system, and takes in a lot of arguments that can update its components.
Half-finished GameEventDisplay, however it needs to be changed such that it works with the current implementation of Sidebar. Just need to implement the cycle pages feature.