Completed GameEventDisplay
GameEventDisplay is now functional. Uses update() function that should be called each iteration. Works on same principle as AttackButtons class: if different information is fed to it, then it updates.
Fixed a bug with mouse_pos - inserted relative_mouse_pos onto AttackButtons and GameEventDisplay button handling, as mouse_pos originally is just determined by the position of the screen.